Hibernate Query Language
Hibernate Query Language (HQL) is a powerful query language provided by Hibernate for querying data from a database using Hibernate entities. It is similar to SQL but operates on the Java objects of the application instead of database tables. Here are some key points and examples related to HQL in your code: